🧹: remove By suffix from UniteBy, DivideBy
Basically Go suffixes func name with "Func" if it accepts a call back function.
And also, we've already name StepBy using By suffix.
Using By to convey that it accepts fn is a bit unnatural for now.
Unite sounds stronger than I want it to be. so let's rename it to Unify.